requirements.txt
pika==1.2.0
SQLAlchemy==1.3.24
pyserial==3.5
psycopg2==2.9.3
psycopg2-binary==2.8.6
pydata-sphinx-theme==0.8.1
